no failure = yes crash

+++ src/main/java/net/voiddustry/redvsblue/player/Hud.java
@@ -57,8 +57,8 @@ public class Hud {
                 RedVsBluePlugin.redSpawns.clear();
                 Groups.build.each(bildeng -> {
                     if (bildeng.block == Blocks.reinforcedLiquidRouter && bildeng.team == Team.all[100]) {
-                        RedVsBluePlugin.redSpawns.add(Vars.world.tile(bildeng.x,bildeng.y));
-                        Vars.world.tile(bildeng.x,bildeng.y).setBlock(Blocks.air);
+                        RedVsBluePlugin.redSpawns.add(Vars.world.tile((int) bildeng.x,(int) bildeng.y));
+                        Vars.world.tile((int) bildeng.x,(int) bildeng.y).setBlock(Blocks.air);
                     }
                 });  
             }
